ml_service/requirements.txt

fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
numpy==1.26.0
opencv-python==4.8.1.78
mediapipe==0.10.8
librosa==0.10.1
tensorflow==2.15.0
motor==3.3.2
pydantic==2.5.2
passlib==1.7.4
python-jose==3.3.0
bcrypt==4.1.1
pytest==7.4.3
httpx==0.25.1
python-dotenv==1.0.0
